# frontend/dist/openapi.yaml
openapi: 3.1.0
info:
  title: Themis API
  version: 0.1.0
  description: |
    Themis network operations API for devices, change requests, approvals, comments,
    golden configs, drift detection, RBAC, AD/LDAP, OAuth/OIDC, and safe deployments.
servers:
  - url: /api
    description: Same-origin Themis API
security:
  - bearerAuth: []
...
paths:
...
  /admin/users/{id}/role:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Admin]
      summary: Change a user's role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: "#/components/schemas/UserRole"
              required: [role]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/settings:
    get:
      tags: [Admin]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      tags: [Admin]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsUpdate"
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserRole:
      type: string
      enum: [viewer, engineer, admin]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, username, email, role]
...
    AppSettings:
      type: object
      additionalProperties: true
      properties:
        max_golden_configs:
          type: integer
        default_required_approvals:
          type: integer
        batch_deploy_concurrency:
          type: integer
        drift_check_interval_secs:
          type: integer
        drift_check_concurrency:
          type: integer
        health_check_concurrency:
          type: integer
        vendor_profiles_toml:
          type: string
        ad_enabled:
          type: boolean
        ad_url:
          type: string
        ad_bind_dn:
          type: string
        ad_bind_password_configured:
          type: boolean
        ad_base_dn:
          type: string
        ad_user_filter:
          type: string
        ad_group_attribute:
          type: string
        ad_default_role:
          $ref: "#/components/schemas/UserRole"
        ad_role_mappings_toml:
          type: string
        oauth_enabled:
          type: boolean
        oauth_provider_name:
          type: string
        oauth_authorize_url:
          type: string
        oauth_token_url:
          type: string
        oauth_userinfo_url:
          type: string
        oauth_client_id:
          type: string
        oauth_client_secret_configured:
          type: boolean
        oauth_redirect_url:
          type: string
        oauth_scopes:
          type: string
        oauth_username_claim:
          type: string
        oauth_email_claim:
          type: string
        oauth_role_claim:
          type: string
        oauth_default_role:
          $ref: "#/components/schemas/UserRole"
        oauth_role_mappings_toml:
          type: string
    AppSettingsUpdate:
      allOf:
        - $ref: "#/components/schemas/AppSettings"
        - type: object
          properties:
            ad_bind_password:
              type: string
              format: password
            oauth_client_secret:
              type: string
              format: password
...